Fix types

+++ include/etl/impl/sse/conv.hpp
@@ -155,9 +155,9 @@ template<typename I, typename K, typename C, cpp_enable_if((all_single_precision
 void conv1_full(const I& input, const K& kernel, C&& conv){
     std::size_t left = size(kernel) - 1;
 
-    double* out = conv.memory_start();
-    const double* in = input.memory_start();
-    const double* k = kernel.memory_start();
+    float* out = conv.memory_start();
+    const float* in = input.memory_start();
+    const float* k = kernel.memory_start();
 
     //Process not-'valid' parts of the convolution (left and right)
     etl::impl::common::left_full_kernel(in, size(input), k, size(kernel), out);
@@ -171,9 +171,9 @@ template<typename I, typename K, typename C, cpp_enable_if((all_single_precision
 void conv1_same(const I& input, const K& kernel, C&& conv){
     std::size_t left = (size(kernel) - 1) / 2;
 
-    double* out = conv.memory_start();
-    const double* in = input.memory_start();
-    const double* k = kernel.memory_start();
+    float* out = conv.memory_start();
+    const float* in = input.memory_start();
+    const float* k = kernel.memory_start();
 
     //Process not-'valid' parts of the convolution (left and right)
     etl::impl::common::left_same_kernel(in, size(input), k, size(kernel), out);
